Magic Squares
This is python code for a genetic algorithm that creates a magic square. Tested for square sizes 3-10. Screenshots of the iterations taken to find a magic square for 5-10 as 3 and 4 are trivial. This includes a Word Doc that explains the logic for the hyper parameters and the genetic operations chosen .